storage: move consistency checker to own file
Pure code movement (mod running goimports). This is in preparation to
recomputing MVCCStats during consistency checks.

Release note: None
