+++ pkg/storage/replica_command.go
@@ -17,14 +17,9 @@ package storage
 import (
 	"bytes"
 	"context"
-	"crypto/sha512"
-	"encoding/binary"
 	"fmt"
-	"io"
 	"math"
-	"sort"
 	"strings"
-	"sync"
 	"sync/atomic"
 	"time"
 
@@ -48,17 +43,6 @@ import (
 	"github.com/cockroachdb/cockroach/pkg/util/log"
 	"github.com/cockroachdb/cockroach/pkg/util/protoutil"
 	"github.com/cockroachdb/cockroach/pkg/util/retry"
-	"github.com/cockroachdb/cockroach/pkg/util/timeutil"
-	"github.com/cockroachdb/cockroach/pkg/util/uuid"
-)
-
-const (
-	// collectChecksumTimeout controls how long we'll wait to collect a checksum
-	// for a CheckConsistency request. We need to bound the time that we wait
-	// because the checksum might never be computed for a replica if that replica
-	// is caught up via a snapshot and never performs the ComputeChecksum
-	// operation.
-	collectChecksumTimeout = 5 * time.Second
 )
 
 // evaluateCommand delegates to the eval method for the given
@@ -697,479 +681,6 @@ func runCommitTrigger(
 	}
 	log.Fatalf(ctx, "unknown commit trigger: %+v", ct)
 	return result.Result{}, nil
-}
-
-// CheckConsistency runs a consistency check on the range. It first applies a
-// ComputeChecksum through Raft and then issues CollectChecksum commands to the
-// other replicas. When an inconsistency is detected and no diff was requested,
-// the consistency check will be re-run to collect a diff, which is then printed
-// before calling `log.Fatal`.
-func (r *Replica) CheckConsistency(
-	ctx context.Context, args roachpb.CheckConsistencyRequest,
-) (roachpb.CheckConsistencyResponse, *roachpb.Error) {
-	desc := r.Desc()
-	key := desc.StartKey.AsRawKey()
-	// Keep the request from crossing the local->global boundary.
-	if bytes.Compare(key, keys.LocalMax) < 0 {
-		key = keys.LocalMax
-	}
-	endKey := desc.EndKey.AsRawKey()
-	id := uuid.MakeV4()
-
-	checkArgs := roachpb.ComputeChecksumRequest{
-		Span: roachpb.Span{
-			Key:    key,
-			EndKey: endKey,
-		},
-		Version:    batcheval.ReplicaChecksumVersion,
-		ChecksumID: id,
-		Snapshot:   args.WithDiff,
-	}
-
-	results, err := r.RunConsistencyCheck(ctx, checkArgs)
-	if err != nil {
-		return roachpb.CheckConsistencyResponse{}, roachpb.NewError(err)
-	}
-
-	var inconsistencyCount int
-
-	for _, result := range results {
-		expResponse := results[0].Response
-		if result.Err != nil || bytes.Equal(expResponse.Checksum, result.Response.Checksum) {
-			continue
-		}
-		inconsistencyCount++
-		var buf bytes.Buffer
-		_, _ = fmt.Fprintf(&buf, "replica %s is inconsistent: expected checksum %x, got %x",
-			result.Replica, expResponse.Checksum, result.Response.Checksum)
-		if expResponse.Snapshot != nil && result.Response.Snapshot != nil {
-			diff := diffRange(expResponse.Snapshot, result.Response.Snapshot)
-			if report := r.store.cfg.TestingKnobs.BadChecksumReportDiff; report != nil {
-				report(r.store.Ident, diff)
-			}
-			buf.WriteByte('\n')
-			_, _ = diff.WriteTo(&buf)
-		}
-		log.Error(ctx, buf.String())
-	}
-
-	// Everything is good, no further action necessary.
-	if inconsistencyCount == 0 {
-		return roachpb.CheckConsistencyResponse{}, nil
-	}
-
-	logFunc := log.Fatalf
-	if p := r.store.TestingKnobs().BadChecksumPanic; p != nil {
-		if !args.WithDiff {
-			// We'll call this recursively with WithDiff==true; let's let that call
-			// be the one to trigger the handler.
-			p(r.store.Ident)
-		}
-		logFunc = log.Errorf
-	}
-
-	// Diff was printed above, so call logFunc with a short message only.
-	if args.WithDiff {
-		logFunc(ctx, "consistency check failed with %d inconsistent replicas", inconsistencyCount)
-		return roachpb.CheckConsistencyResponse{}, nil
-	}
-
-	// No diff was printed, so we want to re-run with diff.
-	// Note that this will call Fatal recursively in `CheckConsistency` (in the code above).
-	log.Errorf(ctx, "consistency check failed with %d inconsistent replicas; fetching details",
-		inconsistencyCount)
-	if err := r.store.db.CheckConsistency(ctx, key, endKey, true /* withDiff */); err != nil {
-		logFunc(ctx, "replica inconsistency detected; could not obtain actual diff: %s", err)
-	}
-
-	// Not reached except in tests.
-	return roachpb.CheckConsistencyResponse{}, nil
-}
-
-// A ConsistencyCheckResult contains the outcome of a CollectChecksum call.
-type ConsistencyCheckResult struct {
-	Replica  roachpb.ReplicaDescriptor
-	Response CollectChecksumResponse
-	Err      error
-}
-
-func (r *Replica) collectChecksumFromReplica(
-	ctx context.Context, replica roachpb.ReplicaDescriptor, id uuid.UUID, checksum []byte,
-) (CollectChecksumResponse, error) {
-	addr, err := r.store.cfg.Transport.resolver(replica.NodeID)
-	if err != nil {
-		return CollectChecksumResponse{}, errors.Wrapf(err, "could not resolve node ID %d",
-			replica.NodeID)
-	}
-	conn, err := r.store.cfg.Transport.rpcContext.GRPCDial(addr.String()).Connect(ctx)
-	if err != nil {
-		return CollectChecksumResponse{},
-			errors.Wrapf(err, "could not dial node ID %d address %s", replica.NodeID, addr)
-	}
-	client := NewConsistencyClient(conn)
-	req := &CollectChecksumRequest{
-		StoreRequestHeader{NodeID: replica.NodeID, StoreID: replica.StoreID},
-		r.RangeID,
-		id,
-		checksum,
-	}
-	resp, err := client.CollectChecksum(ctx, req)
-	if err != nil {
-		return CollectChecksumResponse{}, err
-	}
-	return *resp, nil
-}
-
-// RunConsistencyCheck carries out a round of CheckConsistency/CollectChecksum
-// for the members of this range, returning the results (which it does not act
-// upon). The first result will belong to the local replica.
-func (r *Replica) RunConsistencyCheck(
-	ctx context.Context, req roachpb.ComputeChecksumRequest,
-) ([]ConsistencyCheckResult, error) {
-	// Send a ComputeChecksum which will trigger computation of the checksum on
-	// all replicas.
-	{
-		var b client.Batch
-		b.AddRawRequest(&req)
-
-		if err := r.store.db.Run(ctx, &b); err != nil {
-			return nil, err
-		}
-	}
-
-	var orderedReplicas []roachpb.ReplicaDescriptor
-	{
-		desc := r.Desc()
-		localReplica, err := r.GetReplicaDescriptor()
-		if err != nil {
-			return nil, errors.Wrap(err, "could not get replica descriptor")
-		}
-
-		// Move the local replica to the front (which makes it the "master"
-		// we're comparing against).
-		orderedReplicas = append(orderedReplicas, desc.Replicas...)
-
-		sort.Slice(orderedReplicas, func(i, j int) bool {
-			return orderedReplicas[i] == localReplica
-		})
-	}
-
-	resultCh := make(chan ConsistencyCheckResult, len(orderedReplicas))
-	var results []ConsistencyCheckResult
-	var wg sync.WaitGroup
-
-	for _, replica := range orderedReplicas {
-		wg.Add(1)
-		replica := replica // per-iteration copy for the goroutine
-		if err := r.store.Stopper().RunAsyncTask(ctx, "storage.Replica: checking consistency",
-			func(ctx context.Context) {
-				defer wg.Done()
-
-				ctx, cancel := context.WithTimeout(ctx, collectChecksumTimeout)
-				defer cancel()
-
-				var masterChecksum []byte
-				if len(results) > 0 {
-					masterChecksum = results[0].Response.Checksum
-				}
-				resp, err := r.collectChecksumFromReplica(ctx, replica, req.ChecksumID, masterChecksum)
-				resultCh <- ConsistencyCheckResult{
-					Replica:  replica,
-					Response: resp,
-					Err:      err,
-				}
-			}); err != nil {
-			wg.Done()
-			// If we can't start tasks, the node is likely draining. Just return the error verbatim.
-			return nil, err
-		}
-
-		// Collect the master result eagerly so that we can send a SHA in the
-		// remaining requests (this is used for logging inconsistencies on the
-		// remote nodes only).
-		if len(results) == 0 {
-			wg.Wait()
-			result := <-resultCh
-			if err := result.Err; err != nil {
-				// If we can't compute the local checksum, give up.
-				return nil, errors.Wrap(err, "computing own checksum")
-			}
-			results = append(results, result)
-		}
-	}
-
-	wg.Wait()
-	close(resultCh)
-
-	// Collect the remaining results.
-	for result := range resultCh {
-		results = append(results, result)
-	}
-
-	return results, nil
-}
-
-// getChecksum waits for the result of ComputeChecksum and returns it.
-// It returns false if there is no checksum being computed for the id,
-// or it has already been GCed.
-func (r *Replica) getChecksum(ctx context.Context, id uuid.UUID) (ReplicaChecksum, error) {
-	now := timeutil.Now()
-	r.mu.Lock()
-	r.gcOldChecksumEntriesLocked(now)
-	c, ok := r.mu.checksums[id]
-	if !ok {
-		if d, dOk := ctx.Deadline(); dOk {
-			c.gcTimestamp = d
-		}
-		c.notify = make(chan struct{})
-		r.mu.checksums[id] = c
-	}
-	r.mu.Unlock()
-	// Wait
-	select {
-	case <-r.store.Stopper().ShouldStop():
-		return ReplicaChecksum{},
-			errors.Errorf("store has stopped while waiting for compute checksum (ID = %s)", id)
-	case <-ctx.Done():
-		return ReplicaChecksum{},
-			errors.Wrapf(ctx.Err(), "while waiting for compute checksum (ID = %s)", id)
-	case <-c.notify:
-	}
-	if log.V(1) {
-		log.Infof(ctx, "waited for compute checksum for %s", timeutil.Since(now))
-	}
-	r.mu.RLock()
-	c, ok = r.mu.checksums[id]
-	r.mu.RUnlock()
-	if !ok {
-		return ReplicaChecksum{}, errors.Errorf("no map entry for checksum (ID = %s)", id)
-	}
-	if c.Checksum == nil {
-		return ReplicaChecksum{}, errors.Errorf(
-			"checksum is nil, most likely because the async computation could not be run (ID = %s)", id)
-	}
-	return c, nil
-}
-
-// computeChecksumDone adds the computed checksum, sets a deadline for GCing the
-// checksum, and sends out a notification.
-func (r *Replica) computeChecksumDone(
-	ctx context.Context, id uuid.UUID, sha []byte, snapshot *roachpb.RaftSnapshotData,
-) {
-	r.mu.Lock()
-	defer r.mu.Unlock()
-	if c, ok := r.mu.checksums[id]; ok {
-		c.Checksum = sha
-		c.gcTimestamp = timeutil.Now().Add(batcheval.ReplicaChecksumGCInterval)
-		c.Snapshot = snapshot
-		r.mu.checksums[id] = c
-		// Notify
-		close(c.notify)
-	} else {
-		// ComputeChecksum adds an entry into the map, and the entry can
-		// only be GCed once the gcTimestamp is set above. Something
-		// really bad happened.
-		log.Errorf(ctx, "no map entry for checksum (ID = %s)", id)
-	}
-}
-
-// sha512 computes the SHA512 hash of all the replica data at the snapshot.
-// It will dump all the k:v data into snapshot if it is provided.
-func (r *Replica) sha512(
-	desc roachpb.RangeDescriptor, snap engine.Reader, snapshot *roachpb.RaftSnapshotData,
-) ([]byte, error) {
-	hasher := sha512.New()
-	tombstoneKey := engine.MakeMVCCMetadataKey(keys.RaftTombstoneKey(desc.RangeID))
-
-	// Iterate over all the data in the range.
-	iter := NewReplicaDataIterator(&desc, snap, true /* replicatedOnly */)
-	defer iter.Close()
-
-	var legacyTimestamp hlc.LegacyTimestamp
-	for ; ; iter.Next() {
-		if ok, err := iter.Valid(); err != nil {
-			return nil, err
-		} else if !ok {
-			break
-		}
-		key := iter.Key()
-		if key.Equal(tombstoneKey) {
-			// Skip the tombstone key which is marked as replicated even though it
-			// isn't.
-			//
-			// TODO(peter): Figure out a way to migrate this key to the unreplicated
-			// key space.
-			continue
-		}
-		value := iter.Value()
-
-		if snapshot != nil {
-			// Add the k:v into the debug message.
-			snapshot.KV = append(snapshot.KV, roachpb.RaftSnapshotData_KeyValue{Key: key.Key, Value: value, Timestamp: key.Timestamp})
-		}
-
-		// Encode the length of the key and value.
-		if err := binary.Write(hasher, binary.LittleEndian, int64(len(key.Key))); err != nil {
-			return nil, err
-		}
-		if err := binary.Write(hasher, binary.LittleEndian, int64(len(value))); err != nil {
-			return nil, err
-		}
-		if _, err := hasher.Write(key.Key); err != nil {
-			return nil, err
-		}
-		legacyTimestamp = hlc.LegacyTimestamp(key.Timestamp)
-		timestamp, err := protoutil.Marshal(&legacyTimestamp)
-		if err != nil {
-			return nil, err
-		}
-		if _, err := hasher.Write(timestamp); err != nil {
-			return nil, err
-		}
-		if _, err := hasher.Write(value); err != nil {
-			return nil, err
-		}
-	}
-	sha := make([]byte, 0, sha512.Size)
-	return hasher.Sum(sha), nil
-}
-
-// ReplicaSnapshotDiff is a part of a []ReplicaSnapshotDiff which represents a diff between
-// two replica snapshots. For now it's only a diff between their KV pairs.
-type ReplicaSnapshotDiff struct {
-	// LeaseHolder is set to true of this k:v pair is only present on the lease
-	// holder.
-	LeaseHolder bool
-	Key         roachpb.Key
-	Timestamp   hlc.Timestamp
-	Value       []byte
-}
-
-// ReplicaSnapshotDiffSlice groups multiple ReplicaSnapshotDiff records and
-// exposes a formatting helper.
-type ReplicaSnapshotDiffSlice []ReplicaSnapshotDiff
-
-// WriteTo writes a string representation of itself to the given writer.
-func (rsds ReplicaSnapshotDiffSlice) WriteTo(w io.Writer) (int64, error) {
-	n, err := w.Write([]byte("--- leaseholder\n+++ follower\n"))
-	if err != nil {
-		return 0, err
-	}
-	for _, d := range rsds {
-		prefix := "+"
-		if d.LeaseHolder {
-			// follower (RHS) has something proposer (LHS) does not have
-			prefix = "-"
-		}
-		ts := d.Timestamp
-		const format = `%s%d.%09d,%d %s
-%s  ts:%s
-%s  value:%s
-%s  raw_key:%x raw_value:%x
-`
-		// TODO(tschottdorf): add pretty-printed value. We have the code in
-		// cli/debug.go (printKeyValue).
-		var prettyTime string
-		if d.Timestamp == (hlc.Timestamp{}) {
-			prettyTime = "<zero>"
-		} else {
-			prettyTime = d.Timestamp.GoTime().UTC().String()
-		}
-		num, err := fmt.Fprintf(w, format,
-			prefix, ts.WallTime/1E9, ts.WallTime%1E9, ts.Logical, d.Key,
-			prefix, prettyTime,
-			prefix, d.Value,
-			prefix, d.Key, d.Value)
-		if err != nil {
-			return 0, err
-		}
-		n += num
-	}
-	return int64(n), nil
-}
-
-func (rsds ReplicaSnapshotDiffSlice) String() string {
-	var buf bytes.Buffer
-	_, _ = rsds.WriteTo(&buf)
-	return buf.String()
-}
-
-// diffs the two k:v dumps between the lease holder and the replica.
-func diffRange(l, r *roachpb.RaftSnapshotData) ReplicaSnapshotDiffSlice {
-	if l == nil || r == nil {
-		return nil
-	}
-	var diff []ReplicaSnapshotDiff
-	i, j := 0, 0
-	for {
-		var e, v roachpb.RaftSnapshotData_KeyValue
-		if i < len(l.KV) {
-			e = l.KV[i]
-		}
-		if j < len(r.KV) {
-			v = r.KV[j]
-		}
-
-		addLeaseHolder := func() {
-			diff = append(diff, ReplicaSnapshotDiff{LeaseHolder: true, Key: e.Key, Timestamp: e.Timestamp, Value: e.Value})
-			i++
-		}
-		addReplica := func() {
-			diff = append(diff, ReplicaSnapshotDiff{LeaseHolder: false, Key: v.Key, Timestamp: v.Timestamp, Value: v.Value})
-			j++
-		}
-
-		// Compare keys.
-		var comp int
-		// Check if it has finished traversing over all the lease holder keys.
-		if e.Key == nil {
-			if v.Key == nil {
-				// Done traversing over all the replica keys. Done!
-				break
-			} else {
-				comp = 1
-			}
-		} else {
-			// Check if it has finished traversing over all the replica keys.
-			if v.Key == nil {
-				comp = -1
-			} else {
-				// Both lease holder and replica keys exist. Compare them.
-				comp = bytes.Compare(e.Key, v.Key)
-			}
-		}
-		switch comp {
-		case -1:
-			addLeaseHolder()
-
-		case 0:
-			// Timestamp sorting is weird. Timestamp{} sorts first, the
-			// remainder sort in descending order. See storage/engine/doc.go.
-			if e.Timestamp != v.Timestamp {
-				if e.Timestamp == (hlc.Timestamp{}) {
-					addLeaseHolder()
-				} else if v.Timestamp == (hlc.Timestamp{}) {
-					addReplica()
-				} else if v.Timestamp.Less(e.Timestamp) {
-					addLeaseHolder()
-				} else {
-					addReplica()
-				}
-			} else if !bytes.Equal(e.Value, v.Value) {
-				addLeaseHolder()
-				addReplica()
-			} else {
-				// No diff; skip.
-				i++
-				j++
-			}
-
-		case 1:
-			addReplica()
-
-		}
-	}
-	return diff
 }
 
 // AdminSplit divides the range into into two ranges using args.SplitKey.
